refactor and check all workers are terminated
