+++ services/clusters-keeper/src/simcore_service_clusters_keeper/modules/clusters.py
@@ -55,7 +55,9 @@ async def get_all_clusters(app: FastAPI) -> list[EC2InstanceData]:
     app_settings = get_application_settings(app)
     assert app_settings.CLUSTERS_KEEPER_PRIMARY_EC2_INSTANCES  # nosec
     return await get_ec2_client(app).get_instances(
-        app_settings.CLUSTERS_KEEPER_PRIMARY_EC2_INSTANCES,
+        key_names=[
+            app_settings.CLUSTERS_KEEPER_PRIMARY_EC2_INSTANCES.PRIMARY_EC2_INSTANCES_KEY_NAME
+        ],
         tags=all_created_ec2_instances_filter(app_settings),
         state_names=["running"],
     )
@@ -67,7 +69,9 @@ async def get_cluster(
     app_settings = get_application_settings(app)
     assert app_settings.CLUSTERS_KEEPER_PRIMARY_EC2_INSTANCES  # nosec
     if instances := await get_ec2_client(app).get_instances(
-        app_settings.CLUSTERS_KEEPER_PRIMARY_EC2_INSTANCES,
+        key_names=[
+            app_settings.CLUSTERS_KEEPER_PRIMARY_EC2_INSTANCES.PRIMARY_EC2_INSTANCES_KEY_NAME
+        ],
         tags=ec2_instances_for_user_wallet_filter(
             app_settings, user_id=user_id, wallet_id=wallet_id
         ),
@@ -81,13 +85,14 @@ async def get_cluster_workers(
     app: FastAPI, *, user_id: UserID, wallet_id: WalletID | None
 ) -> list[EC2InstanceData]:
     app_settings = get_application_settings(app)
-    assert app_settings.CLUSTERS_KEEPER_PRIMARY_EC2_INSTANCES  # nosec
+    assert app_settings.CLUSTERS_KEEPER_WORKERS_EC2_INSTANCES  # nosec
     return await get_ec2_client(app).get_instances(
-        app_settings.CLUSTERS_KEEPER_PRIMARY_EC2_INSTANCES,
+        key_names=[
+            app_settings.CLUSTERS_KEEPER_WORKERS_EC2_INSTANCES.WORKERS_EC2_INSTANCES_KEY_NAME
+        ],
         tags={
             "Name": f"{get_cluster_name(app_settings, user_id=user_id, wallet_id=wallet_id, is_manager=False)}*"
         },
-        state_names=["pending", "running"],
     )
