+++ services/clusters-keeper/tests/unit/conftest.py
@@ -22,6 +22,8 @@ from asgi_lifespan import LifespanManager
 from faker import Faker
 from fakeredis.aioredis import FakeRedis
 from fastapi import FastAPI
+from models_library.users import UserID
+from models_library.wallets import WalletID
 from moto.server import ThreadedMotoServer
 from pytest_mock.plugin import MockerFixture
 from pytest_simcore.helpers.utils_docker import get_localhost_ip
@@ -37,6 +39,7 @@ from simcore_service_clusters_keeper.modules.ec2 import (
     ClustersKeeperEC2,
     EC2InstanceData,
 )
+from simcore_service_clusters_keeper.utils.ec2 import get_cluster_name
 from types_aiobotocore_ec2.client import EC2Client
 from types_aiobotocore_ec2.literals import InstanceTypeType
 
@@ -407,3 +410,66 @@ async def clusters_keeper_rabbitmq_rpc_client(
     rpc_client = await rabbitmq_rpc_client("pytest_clusters_keeper_rpc_client")
     assert rpc_client
     return rpc_client
+
+
+@pytest.fixture
+def create_ec2_workers(
+    aws_ami_id: str,
+    ec2_client: EC2Client,
+    user_id: UserID,
+    wallet_id: WalletID,
+    app_settings: ApplicationSettings,
+) -> Callable[[int], Awaitable[list[str]]]:
+    async def _do(num: int) -> list[str]:
+        instance_type: InstanceTypeType = "c3.8xlarge"
+        assert app_settings.CLUSTERS_KEEPER_WORKERS_EC2_INSTANCES
+        instances = await ec2_client.run_instances(
+            ImageId=aws_ami_id,
+            MinCount=num,
+            MaxCount=num,
+            InstanceType=instance_type,
+            KeyName=app_settings.CLUSTERS_KEEPER_WORKERS_EC2_INSTANCES.WORKERS_EC2_INSTANCES_KEY_NAME,
+            TagSpecifications=[
+                {
+                    "ResourceType": "instance",
+                    "Tags": [
+                        {
+                            "Key": "Name",
+                            "Value": f"{get_cluster_name(app_settings,user_id=user_id,wallet_id=wallet_id,is_manager=False)}_blahblah",
+                        }
+                    ],
+                }
+            ],
+        )
+        print(f"--> created {num} new instances of {instance_type=}")
+        instance_ids = [
+            i["InstanceId"] for i in instances["Instances"] if "InstanceId" in i
+        ]
+        waiter = ec2_client.get_waiter("instance_exists")
+        await waiter.wait(InstanceIds=instance_ids)
+        instances = await ec2_client.describe_instances(InstanceIds=instance_ids)
+        assert "Reservations" in instances
+        assert instances["Reservations"]
+        assert "Instances" in instances["Reservations"][0]
+        assert len(instances["Reservations"][0]["Instances"]) == num
+        for instance in instances["Reservations"][0]["Instances"]:
+            assert "State" in instance
+            assert "Name" in instance["State"]
+            assert instance["State"]["Name"] == "running"
+            assert "Tags" in instance
+            for tags in instance["Tags"]:
+                assert "Key" in tags
+                if "Name" in tags["Key"]:
+                    assert "Value" in tags
+                    assert (
+                        get_cluster_name(
+                            app_settings,
+                            user_id=user_id,
+                            wallet_id=wallet_id,
+                            is_manager=False,
+                        )
+                        in tags["Value"]
+                    )
+        return instance_ids
+
+    return _do
